+++ src/main/java/com/example/Lion.java
@@ -6,7 +6,8 @@ public class Lion {
 
     boolean hasMane;
 
-    public Lion(String sex) throws Exception {
+    public Lion(String sex, Feline feline ) throws Exception {
+        this.feline = feline;
         if ("Самец".equals(sex)) {
             hasMane = true;
         } else if ("Самка".equals(sex)) {
@@ -17,11 +18,8 @@ public class Lion {
     }
 
 
+     private Feline feline;
 
-    public Lion(Feline feline) {
-        this.feline = feline;
-    }
-     Feline feline;
     public int getKittens() {
         return feline.getKittens();
     }
