+++ src/test/java/com/example/FelineTest.java
@@ -2,23 +2,15 @@ package com.example;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.MockitoJUnitRunner;
 import java.util.List;
 
 import static org.junit.Assert.*;
-@RunWith(MockitoJUnitRunner.class)
 public class FelineTest {
-    @Mock
-    Animal animal;
 
     @Test
         public void checkEatMeatForFeline() throws Exception {
             Feline feline = new Feline();
-            Mockito.when(animal.getFood("Хищник")).thenReturn(List.of("Животные", "Птицы", "Рыба"));
-            feline.eatMeat();
+
             List<String> expected = List.of("Животные", "Птицы", "Рыба");
             List<String> actual = feline.eatMeat();
             Assert.assertEquals(expected, actual);
@@ -27,8 +19,8 @@ public class FelineTest {
     @Test
     public void checkGetFamilyForFeline() {
         Feline feline = new Feline();
-        String actual = "Кошачьи";
-        String expected = feline.getFamily();
+        String actual = feline.getFamily();
+        String expected = "Кошачьи";
         assertEquals(expected, actual);
     }
